feat: add credit card entity
