+++ internal/enums/card_color.enum.go
@@ -1,0 +1,10 @@
+package enums
+
+type CardColor string
+
+const (
+	LIGHT_BLUE     CardColor = "LIGHT BLUE"
+	BLUE           CardColor = "BLUE"
+	DARK_BLUE      CardColor = "DARK BLUE"
+	VERY_DARK_BLUE CardColor = "VERY DARK BLUE"
+)
